+++ tower-defense/src/game_state/hand/mod.rs
@@ -1,6 +1,7 @@
 mod hand_slot;
 mod render_card;
 mod render_tower;
+mod shared;
 mod xy_with_spring;
 
 use crate::{
