+++ tower-defense/src/game_state/hand/render_card.rs
@@ -2,10 +2,10 @@ use crate::{
     card::{Card, Rank},
     icon::{Icon, IconKind, IconSize},
     theme::{
-        palette,
         typography::{FontSize, TextAlign, headline},
     },
 };
+use super::shared::{render_background_rect, render_top_left_rank_and_suit};
 use namui::*;
 
 pub(super) struct RenderCard {
@@ -18,7 +18,7 @@ impl Component for RenderCard {
         let Self { wh, card } = self;
 
         // 좌상단에 숫자와 문양 수직 배치
-        self.render_top_left_rank_and_suit(ctx, wh, card);
+        render_top_left_rank_and_suit(ctx, card.rank, card.suit);
 
         // 중앙에 알맞은 수의 문양 배치 (숫자 카드만)
         if !card.rank.is_face() {
@@ -28,49 +28,11 @@ impl Component for RenderCard {
             self.render_face_card_placeholder(ctx, wh);
         }
 
-        ctx.add(rect(RectParam {
-            rect: wh.to_rect(),
-            style: RectStyle {
-                stroke: Some(RectStroke {
-                    color: palette::OUTLINE,
-                    width: 4.px(),
-                    border_position: BorderPosition::Inside,
-                }),
-                fill: Some(RectFill {
-                    color: palette::SURFACE_CONTAINER,
-                }),
-                round: Some(RectRound {
-                    radius: palette::ROUND,
-                }),
-            },
-        }));
+        render_background_rect(ctx, wh);
     }
 }
 
 impl RenderCard {
-    fn render_top_left_rank_and_suit(&self, ctx: &RenderCtx, _wh: Wh<Px>, card: Card) {
-        let padding = px(4.0);
-        let rank_font_size = FontSize::Small;
-        let suit_icon_size = px(16.0);
-
-        // 숫자 렌더링
-        ctx.translate(Xy::new(padding, padding)).add(
-            headline(card.rank.to_string())
-                .size(rank_font_size)
-                .align(TextAlign::LeftTop)
-                .build(),
-        );
-
-        // 문양 아이콘 렌더링 (숫자 아래)
-        ctx.translate(Xy::new(padding, padding + px(20.0))).add(
-            Icon::new(IconKind::Suit { suit: card.suit })
-                .wh(Wh::new(suit_icon_size, suit_icon_size))
-                .size(IconSize::Custom {
-                    size: suit_icon_size,
-                }),
-        );
-    }
-
     fn render_center_suits(&self, ctx: &RenderCtx, wh: Wh<Px>, card: Card) {
         let center_area = Rect::Xywh {
             x: px(36.0),
